+++ src/commands.rs
@@ -1,6 +1,7 @@
 mod convert;
 mod cp;
 mod down;
+mod events;
 mod pause;
 mod ps;
 mod rm;
@@ -24,6 +25,7 @@ pub(crate) enum Command {
     Pause(pause::Args),
     Unpause(unpause::Args),
     Cp(cp::Args),
+    Events(events::Args),
     Top(top::Args),
     Ps(ps::Args),
     Convert(convert::Args),
@@ -39,6 +41,7 @@ pub(crate) async fn run(command: Command, config: Config) -> Result<()> {
         Command::Pause(args) => pause::run(args, config).await?,
         Command::Unpause(args) => unpause::run(args, config).await?,
         Command::Cp(args) => cp::run(args, config).await?,
+        Command::Events(args) => events::run(args, config).await?,
         Command::Top(args) => top::run(args, config).await?,
         Command::Ps(args) => ps::run(args, config).await?,
         Command::Convert(args) => convert::run(args, config)?,